[cv][cqd] avoid confusion in migration API's Run message.
Before, the message was used for 2 different purposes.
Now, there are 2 messages to avoid confusion about which are set
and which are not: ReportedRun to be populated by CQDaemon for CV's
consumption and ActiveRun populated by CV for CQDaemon.

This requires changes on CQDaemon side, so rolls won't be trivial.

However, the canary/prod CQDaemon will still be able to read
FetchActiveRunsResponse, but it won't see any active_runs there,
which uses a different field number currently not known to CQDaemon.

Also, because ReportedRun uses exact same field numbers as Run message,
CV will continue being able to correctly handle RPCs like
ReportFinishedRun.

R=qyearsley, yiwzhang

Bug: 1160019
Change-Id: I3f3c2b53663feb056e18f55b2c6db69bc2a23c4d