[cv][cqdfake] don't block ActiveAttemptKeys on internal mutex.
The CQDfake background goroutine, which simulates CQDaemon main thread,
may be blocked on RPC to CV Migration API,
which in turn may be trying to schedule a task via TQ,
which in tests is done via tqtesting which has a single mutex,
which in turn may be blocked on *callback* checking stop condition,
which in turn may be executing e2e's RunUntil's callback,
which in turn may be checking the active attempts of CQDfake,
causing deadlock.

The easiest workaround is to ensure CQDfake state examining functions
aren't interfering with CQDFake's own work. It's also how real CQDaemon
operates.

R=qyearsley, yiwzhang

Bug: 1160019
Change-Id: I042fa5e9c2f841648e4c0c77c147dcb601672fac