[buildbucket] Refactor build proto creation to support dry runs
In the future, we want to be able to get the Build proto output of a
ScheduleBuild request without performing any mutative operations.

As part of this, rather than having setExecutable and setInfra rely on
an experiment value to set the command and priority respectively, just
set those values during setExperiments. This allows preparing most of
the proto independently of setting values in the entity which will
support a future dry run flow in which the proto is just returned
without actually scheduling the build.

Bug: 1042991
Change-Id: I83a76873b554d9b711e3640949499c4fbb859379