cv: Implement CL Group Key functions for BQ export
Notes:

The information used in the hash is the same as for CQD:
The Gerrit host, CL number and patchset (or min equivalent patchset).

For reference, CQD cl group key digest is computed as:

    triples = sorted((cl.hostname, cl.number, cl.patchset) for cl in self.cls)
    key = '\x00'.join(str(x) for x in itertools.chain.from_iterable(triples))
    return hashlib.sha224(key).hexdigest()

Since either EquivalentClGroupKey or ClGroupKey is changing anyway
so that they're not equal with each other, and because potential
the breakage scenario is tiny (it would only affect rows for the
same patchset combo during the time of migration) I think it's OK
to change the way this is computed a little.

Bug: 1173168
Change-Id: Iea0e3a6bd7ba9e1f1902118f6aa04ac797af33c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2894693
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>