cv: move acquiring submit queue and tree check logic to OnReadyForSubmission
It is a bug that those logic are not in OnReadyForSubmission handler
because Run will be submitted without acquiring the Submit Queue(SQ). I
think I implemented this way with the assumption that when
OnReadyForSubmission event is received, Run is guaranteed to hold the
SQ. However, I changed this behavior later on but didn't go back and
modify this logic.

Also, in this change, tree check is now performed after successfully
acquiring the SQ. If tree is closed, SQ will be released to unblock
other Runs in the same Project from submitting while waiting for the
tree to re-open.


R=qyearsley, tandrii

Bug: 1179284
Change-Id: I762f4005d4e2fef140d7e7ed865936abfcd787b1