Use os.Rename for atomic move on Windows
Previously, CIPD used a custom implementation of file moving
on Windows that syscall-ed to `MoveFileEx` in order to approximate
atomic filesystem operations.

Of particular interest in the implementation is the use of the
flag `MOVEFILE_WRITE_THROUGH`. In the last few years the
documentation of this flag was clarified:

> The function does not return until the file is actually moved
> on the disk. Setting this value guarantees that a move
> performed as a copy and delete operation is flushed to disk
> before the function returns. The flush occurs at the end of the
> copy operation.

However, the only case where a move can be performed as a
copy and delete operation is if the flag `MOVEFILE_COPY_ALLOWED`
is also specified:

> If the file is to be moved to a different volume, the function
> simulates the move by using the CopyFile and DeleteFile functions.
> If the file is successfully copied to a different volume and the
> original file is unable to be deleted, the function succeeds
> leaving the source file intact.

https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-movefileexw

Because the current implementation does not specify this flag,
moves across volumes would currently fail. Furthermore, copies
are unlikely to be across volumes since atomic moves are part of
the process to deploy packages from the `<root>/.cipd` to
`<root>/src` directories.

Lastly, I assume that the vast majority of CIPD users are running
on the NTFS filesystem (as opposed to something like FAT). NTFS
has a volume-wide journal, so any filesystem operations that are
dependent on the move operation won't be durable unless the move
itself succeeds.

As a result, the MoveFileEx call does the equivalent of a POSIX
`fsync()` operation for every file move, which appears to be
overkill for our durability requirements (and is more pessimistic
than the POSIX implementation of `mostlyAtomicRename`).

While I could simply remove that flag from CIPD's syscall, doing
so would bring the implementation in line with golang's (new as
of 2015) built-in implementation of `os.Rename` for Windows
https://go-review.googlesource.com/c/go/+/6140/. From what I can
see it appears that this CIPD implementation likely predates this
standard library change, and as seen in the diff, the old
implementation of os.Rename failed when attempting to overwrite
files. As a result, I replaced the syscall with the standard libary
call instead to reduce the maintenance burden.

Change-Id: I9c02ed54a91cc5760001305b272c123f52a6947e