[cv][itriager] implement triageComponents and actOnComponents.
The ideas of saving partial progress and handling panics are
borrowed from existing functions in the same file above/below
the modifications (for ease of reviewers), but I hope new code
is cleaner. Furthermore, the tests cover run creation,
s.t. Triager implementations are easier to test.

Old functions are preserved until componentActor is refactored
to be a triager.

R=qyearsley, yiwzhang

Bug: 1156241
Change-Id: Ibcd0859e0b87512f5fe679b6b1829ffb4d274d67