cv: make OnReadyForSubmission noop if run is submitting.
I was trying to be defensive at the beginning to make sure submission
can make progress even if it is halted for whatever reason. But now
another handler will always be invoked at the end of state transition
to ensure that, so there's no need to handle in OnReadyForSubmission
handler anymore.

R=qyearsley, tandrii

Bug: 1179284
Change-Id: Ie9f24c06c70730557ac8af70c5e2b8bc7e446fb0