[cv] start refactoring to get rid of "actor" struct.
The actor struct makes no sense any more, and it'd better to pass
state between functions explicitly for readability of the data flow.

Of course, there will still be a helper object when doing a run creation.

This CL removes computation of reverseDeps as part of CL triage,
as in its current form it's not useful. reverseDeps was originally
planned for multi-CL Run creation, but I've learned yesterday that
i'll need a slightly different map. So, I'll add its computation right
before it is needed in the follow up CLs.

R=qyearsley, yiwzhang

Bug: 1156241
Change-Id: I9646bea4cca1a88cccfa6d95a99abe66d79c80de
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2912108
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>