[cv][gerrit] spread out load on Gerrit during full poll.
Currently, every ~30 mins CV force-refreshes all CLs which used to be
returned by Gerrit for a CQ query but no longer are.
This catches updates which somehow didn't make it into Gerrit index OR
due to eventual consistency weren't visible to CQ.

In some cases there can be LOTs of such CLs, e.g. due to config changes
which make lots of previously watched CLs unwatched. See crbug/1211057
for an example.

So, distribute the refreshes over the 30 minute period.

R=qyearsley, yiwzhang

Bug: 1211057, 1181978
Change-Id: I95f0fc4da2e8ea123e6fc7a79e8f77370ad9fb6b
Fixes: 1181978