[resultdb] Gen paging predicate conditionally
When generating the SQL query for fetch artifacts, include current page
filter only if the page token was specified.
This is consistent with test result query.
It might improve the performance because the expression is quite
complicated and Spanner might be not smart enough to remove it if it
page token isn't specified.

Bug: 1215246
Change-Id: I63b979eaf3508a6ea4cd3028014ce52391c66313