[cv] partial fix for missing CL updated notifications in PM.
This isn't completely correct: in the future, there can be several
incomplete Run per CL, and each of them have to be notified. This code
currently assumes at most 1 Run per CL, so notifying it isn't critical.

This is also not ideal, as PM is now awoken right after Run creation,
which is somewhat wasteful.

Last but not least, this CL adds regression tests, allowing for
relatively safe refactoring in the future.

R=qyearsley, yiwzhang

Bug: 1215792
Change-Id: I84f34c80be9dab0a39ed6af36dccda0da6e27fa2