Fix "com.docker.cli" not found in $PATH
This happens when running tests with the spanner emulator in Mac.
The previously available workaround (disabling 'cloud experience' in
Docker Desktop) no longer works as of docker desktop 3.3.3

I found that allowing the cmd running `docker run` have access to PATH
fixes it for me.

My understanding is that `docker` now requires that `com.docker.cli` be
findable in its PATH, but exec.Command() only resolves the path to
`docker` and does not leak the whole contents of PATH to the process.

R=chanli,ddoman,tandrii

Change-Id: I0560cd150545e2f141312d9cc4758d5a4d0b765d