[resultdb] Fix race in test of deadline enforcer.
For test purposes we can enforce each shard exactly once, and know
with certainty that the function has finished before checking the
results.

In the previous approach we would run the enforcer asynchronously and
wait on one of its effects, but assume that it had finished.

R=tandrii,ddoman,chanli
Bug: 1210505
Change-Id: I275bbd9b1723fd54c98f5c77dda6b7932ac42bcd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2941542
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>