[cv][pm][rm] handle Datastore cross-transaction contention.
Handle it the same way as changed EVersion was done previously.
The root cause is the same, but the way it comes about is slightly
different.

Eversion changed means:
  handler A: reads entity
  handler A: does processing
  handler B: reads entity
  handler A: starts transaction on entity
  handler A: entity.EVersion++
  handler A: commits transaction
  handler B: does processing
  handler B: starts transaction on entity
  handler B: aborts transaction because EVersion changed.

Datastore cross-transaction contention is like this:
(though not not only limited to this):
  handler A: reads entity
  handler A: does additional processing
  handler B: reads entity
  handler B: does additional processing
  handler A: starts transaction on entity
  handler B: starts transaction on entity
  handler A: entity.EVersion++
  handler B: **Datastore** aborts B's transaction
  handler A: commits transaction

For the purposes of eventbox, and hence Run and Project Managers,
these are ultimately due to the same root cause: concurrent
invocations of Run or Project managers.
So, they are now treated the same way.

Also, move logging.SetField right into task handler because this is
where the final manage() error is logged.

R=qyearsley, yiwzhang

Change-Id: I178308acffcbf14350719cda29f978c7dd404928
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2953383
Auto-Submit: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>