[cv][eventbox] s/Mutate/PrepareMutation.
I think this is more accurate because Mutate() isn't run in a
transaction, so it's more of a factory that produces
"callMeToMutate" functions which are later ran inside a transaction.

However, in defense of old Mutate() name, the Mutate() may
have some immediate effects on the CV state, such as newly created
Runs (in PM) or acquired submission queue (in RM).

Overall, I think PrepareMutation is more appropriate.

R=qyearsley, yiwzhang

Change-Id: I89ce9968e8d0ecef5535bcb041ebbe6728c47401
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2953384
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>