[cipd] Make sure EnsurePackages always has an instance cache.
To simplify implementation of parallel package downloads it is
beneficial to assume we always have some instance cache available,
so all fetching can be done there. It reduces branching on
CIPD_CACHE_DIR availability.

So now if CIPD_CACHE_DIR is empty, the client will construct
a temporary cache directory and use it instead of the global one.
It has a property that files there "self-destruct" after being
consumed, and the entire directory is nuked when the client is
done running `cipd ensure`.

Such temporary cache can be understood as "a cache of downloaded,
but not yet installed, files" (rather than just "a cache of
downloaded files").

R=iannucci@chromium.org
BUG=1217435

Change-Id: I659b20d703d3c1545f89920aebc62e7ba759fbf6