[cv] reduce Datastore contention on the busiest LUCI projects.
Previously, if Datastore contention was detected, the TQ task will
be retried, which for busiest projects is likely to be concurrent
with another newly scheduled task, which leads to continued contention.

Now, if Datastore contention is detected, a new TQ task is created,
which provides for de-duplication, while the current task is failed
permanently. This should reduce chances of future contention.

R=yiwzhang

Change-Id: Ic3389627479e9939bb496dbfad8c09d2c058cd1a