[cipd] Refactor EnsurePackages to support parallel fetches.
The focus of this CL is changes to client.go and to
the InstanceCache API it uses. InstanceCache implementation is
currently a complete mess and will be refactored in a later CL.

Note that the parallelism level is currently hardcoded to 1 (since
the code that actually implements it is not really ready yet) and
as a result this should be a pure refactoring with no (significant)
observable differences. In particular, packages are still being
unzipped in the order of their appearance in the manifest file.

R=iannucci@chromium.org
BUG=1217435

Change-Id: Icc09c690fe0fa149577619b0c3cf3aeb15ccfff5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2971199
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>