[cipd] Redo how progress logging is done.
All extended activities (uploads, downloads, unzipping) are now
using a uniform way to report their progress. Each such operation
is represented by an ui.Activity that gets progress reports and logs
from a particular serial process (using the context magic).

This would allow the cipd CLI to render a nice UI for them with
fancy terminal progress bars and what not (haven't tried yet).

For cases when the client is used as a library or when cipd CLI is
not attached to a terminal, fall back to multiplexed logging. For now
it the only available UI implementation.

Drop silly "cipd: ..." logging prefixes in client.go and storage.go.
They'll be eventually dropped from other places too.

R=iannucci@chromium.org
BUG=1217435

Change-Id: Ic785bd313a3db2ca2ccd2b608e5811de160ccace