[buildbucket] Add legacy field to buckets
Go cannot compute this because proto fields are non-pointer, so the
bucket proto's Swarming field is always set. None of the fields within
are validated by Python as required, so there's no way for Go to
differentiate an empty Swarming field from no Swarming field. This field
is set by Python as a hint for Go.

Bug: 1042991
Change-Id: Ia4ef7c77f20c055f6c7ce32b6ddc0684a7e3c505