[cv][gerrit] instrument the client with grpcmon-like metrics.
Unlike gRPCmon, these include LUCI project & gerrit host.

The main downside is that there is a risk that this will have high
cardinality if we experience too many grpc codes in responses.

I think it's worth having this in the short term until we can
take advantage of custom targets.

R=yiwzhang

Change-Id: Ie9895de2298ebb115ec04dfd7842578a68768724
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3003436
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>