[go] Fixes Starlark tests in Go modules mode.
go.starlark.net/starlarktest needs assert.star file from within
its package code, but getting a path to it is non-trivial in Go
modules mode. Instead transform it into a *.go file and materialize
it on disk on demand in a temp directory.

go.starlark.net/starlarktest has a TODO to switch to "go embed"
eventually. Once it happens, this hack can be removed.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1057067

Change-Id: If2abdf44ad2995ee0655d2395d318ceeb74414ba