[cv][cl] use CL Mutator when creating entities for CL dependencies.
This has a minor of downside of notifying PM every time a CL is created.
However, it's not critical for performance, and consistency of all
mutations going via CL Mutator is a bigger win.

ExternalID.GetOrInsert is simplified and kept for use in tests.

R=yiwzhang

Change-Id: I26c324a0b752258a931d8ba5984d879246123dc8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3010109
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>