Reland "[cv][gerrit] iterate over Gerrit mirrors if stale data is detected."
This is a reland of 61a875790fbe1a142dedeb167742833f0ed3b513
Includes fixes and simplification.

The root problem before was that *MirrorIterator was considered
* a grpc.CallOption by Go type system,
* not an opaque type which gerrit.UseGerritMirror(...) returned

Unfortunately, due to use of the opaque type, it's not trivial to test:
either I need the opaque type to become public in gerrit package, OR,
to construct new gerrit.UseGerritMirror in each iteration.

I chose the latter because it provides for (future work) "freezing"
mirror iterator at an up-to-date mirror and then using this mirror
for subsequent Gerrit RPCs when fetching change details.

Original change's description:
> [cv][gerrit] iterate over Gerrit mirrors if stale data is detected.
>
> This CL doesn't yet configure any mirrors, so this alone should have
> no effect on production.
>
> R=ddoman, yiwzhang
>
> Bug: 1220461
> Change-Id: I3545cc2ae5ec440a835f3238b459850a02eaff65
> Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3021224
> Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
> Reviewed-by: Yiwei Zhang <yiwzhang@google.com>

Bug: 1220461
Change-Id: Ie72a84bbc34407b398d1d6d2eb219d84a38068cb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3024287
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>