resultdb: add failure_reason to test_result
failure_reason allows capturing the reason for failing tests. This
information is initially slated for use in clustering failures by
Weetbix.

This CL adds the change to the db schema and resultdb services.  Follow
up CLs will modify sink and test frameworks to actually provide the
failure reasons.

I believe the database schema will need to be updated before this CL can
be landed, I'm not sure of the process for this.  There is minimal risk
to the DB from this change as it is simply adding a new unindexed
column.

Bug: 1226947
Change-Id: Ib32d9d923935b5702e6b0d2e66e19043cfb96c63
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3010780
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>