[led] Add a -property-only flag to edit-recipe-bundle
For builders that use bbagent with a "bootstrap" executable that
executes a recipe under the hood, it's not possible to do `led
edit-recipe-bundle` without overwriting the executable, causing the led
job to run the recipe directly, rather than indirectly via the bootstrap
executable. That breaks any recipe logic that assumes the recipe is run
by a specific executable.

To work around that, this change adds a `-property-only` flag to the
`edit-recipe-bundle` command. When the flag is set, `edit-recipe-bundle`
will still bundle and upload the local recipes but will not override the
executable used by the build. Instead, led will pass the the CAS
reference of the recipe bundle as a property for the job's exe to read.

In the case of a bootstrap executable like fuchsia uses, this will
require a corresponding change to check if the `led_cas_recipe_bundle`
property is set, and if so, download and execute that recipe bundle
rather than following the normal recipe checkout process.

Test: `led get-build 8841535313446777120 | led edit-recipe-bundle -property-only | led launch`
https://ci.chromium.org/swarming/task/54ee3a7b4b3a5510?server=chromium-swarm.appspot.com

Bug: 1216487
Change-Id: Ia2d929bc6d3487b58a55a4b58e917cfdf3986fb0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3035428
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>