Use GCE's zone and instance name on GCE.
When we started to use
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3023167,
stackdriver started to reject our requests with Invalid argument error.
We suppose that stackdriver won't accept non-existing zone.

Also, as far as I understand the opencensus library, the zone parameter
looks optional.  If the zone is not returned by the GCE metadata server,
the library warns but it won't be handled as error.
https://github.com/census-ecosystem/opencensus-go-exporter-stackdriver/blob/master/stackdriver.go#L328

Let me make the fake GCE metadata server returns zone and name only if
these values are available on the host environment.

Bug: 1219914
Change-Id: I4111e30bb0cc2edb226e98c457cd9f9f519cc5f8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3057313
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>