[server/tq] add a new RunOption, SkipPubSubTask, in tqtesting
In CV, a Run termination enqueues both TaskQueue and PubSub
tasks.
- bq-export and pubsub for run-events.

However, the executor in server/tq/testing.go panics if it receives
a PubSub task, and causes tests that invokes tqtesting/Scheduler:Run()
to panic.

This CL adds a new option, SkipPubSubTask, and it can be used to
tell the scheduler to skip executing PubSub tasks, if set.

Bug: 1226144
Change-Id: I3c66b6abf2e461d281a4d32b55d37545937ab403
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3094605
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>