Add support for a property to override edit-recipe-bundle behavior.
This changes causes the edit-recipe-bundle command to look for the
led_builder_is_bootstraped property. If it's set and the value is true,
the CAS digest will be written to the led_cas_recipe_bundle property
instead of overriding the build's task payload.

This enables overriding the recipe for a builder using a bootstrapper
executable without forcing the led user to know whether or not the
builder is bootstrapped. Using lucicfg, it is possible to have the
led_builder_is_bootstrapped property automatically declared to the
appropriate value for builders that make use of a bootstrapper
executable, making the presence of a bootstrapper binary transparent to
the user of led.

Test: led get-build 8838535025622636816 | led edit -p led_builder_is_bootstrapped='true' | led edit-recipe-bundle | led launch
https://ci.chromium.org/swarming/task/5574186a989d6311?server=chromium-swarm.appspot.com

Change-Id: Ib3d3f4ab82e9e6f4282ae54b078aed93544e6e03
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3104295
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>