[testclock] de-racify FastClock tests.
The root problem is with the FastClock.Set call:
because it's a clock which is constantly moving forward,
the following may cause panic:
  t := FastClock.Now()
  t.Add(time.Second)
  FastClock.Set(t)  // panics if process was starved of CPU for >1 second.

The test itself doesn't strictly need this, so I rewrote the test.
However, it may be a good idea to make FastClock.Set ignore time which
is already in the past. It still makes it possible to use in-place
of where testClock was used before, but it's now guaranteed not to panic.

R=vadimsh, yiwzhang

Change-Id: I58bf406acae0d99250410b72b117358ca03a166d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3107125
Auto-Submit: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>