[datastore] intelligently zero out fields when loading from Datastore.
Excludes magic $id & $parent, which are usually set by the user
and may be used by the user when triaging/diagnosing decoding errors.

This is a not a backwards compatible change. Notably, I can think of
two possible usages that are affected:

1. Gotcha with repeated fields, which is unlikely to be relied upon intentionally:

    type MyObj struct {
      ID     string   `gae:$id`
      Labels []string
    }

    o := MyObj{ID: "blah", Labels:[]string{"a"}
    _ = datastore.Put(ctx, &o)
    _ = datastore.Get(ctx, &o)
    // BEFORE: o.Labels == []string{"a", "a"}.
    // AFTER:  o.Labels == []string{"a"}.


2. Setting default fields values before reading from Datastore,
possibly for backwards compatibility. For example,

    // OldObj is how objects were written historically.
    type OldObj struct {
      _kind   string   `gae:"$kind,Obj"`
      ID      string   `gae:$id`
      Field   string
    }
    // NewObj is the new layout, which is backwards compatible.
    type NewObj struct {
      _kind   string   `gae:"$kind,Obj"`
      ID      string   `gae:$id`
      Field   string
      Version string   // NOTE: old objects don't have this field set.
    }

    // Loads an object. Guarantees that its .Version is set.
    func Loads(ctx context.Context, id string) *NewObj {
      ret := &NewObj{
        ID:      id,
        Version: "v0", // datastore.Get will set overwrite this if available.
      }
      _ = datastore.Get(ctx, ret)
      return ret
    }

    _ = datastore.Put(ctx, &OldObj{ID: "legacy", Field: "foo"})
    _ = datastore.Put(ctx, &NewObj{ID: "modern", Field: "bar", Version: "v1"})

    modern := datastore.Get(ctx, "modern")
    fmt.Println(modern.Version)  // outputs "v1"

    legacy := datastore.Get(ctx, "legacy")
    fmt.Println(legacy.Version)
    // BEFORE: outputs "v1"
    // AFTER:  outputs ""


I'm not sure if use case of (2) can be ignored, nor do I know how this library
can detect this. If we are lucky, the users of (2) have good test cases,
but such test case isn't trivial to write, so I doubt we are lucky.

Therefore, this functionality is currently opt in via datastore.EnableSafeGet(),
introduced in prior CL.
After ~4 month or so, I hope the error logs will nudge anyone who cares about
their app to opt in and deploy a fix.

R=iannucci

Bug: 1189937, 1242998
Change-Id: If8bb911b413e9ddb09fe0785c36121dc4346c725
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3103793
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>