[auth-service] Add a workaround for broken RPC Explorer.
RPC Explorer is that thing:
https://defaultv2-dot-chrome-infra-auth-dev.appspot.com/rpcexplorer/

It uses OAuth2 access tokens to authenticate calls to the backend.
Unfortunately it also uses XMLHttpRequest to actually makes requests,
and they end up with all cookies, since this is the hardcoded logic
of XMLHttpRequest.

The RPC Explorer is totally unaware of cookies. It uses OAuth2 tokens
in "Authorization" header. When it sends a request without
"Authorization" header, it expects this request to be executed
anonymously. Instead browser also silently sends cookies, which
engage `srv.CookieAuth` authenticator. Even if cookies are valid,
xsrf.Interceptor(...) then rejects the request, since it doesn't
have X-Xsrf-Token header.

As a workaround we just completely ignore cookies if X-Xsrf-Token
header is absent.

R=cjacomet@google.com, tandrii@google.com
BUG=1112832

Change-Id: I4e51a4d2dec6e0ea4cef387be9a6159d62420de6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3126685
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Camilo Jacomet <cjacomet@google.com>