[cv] Record queue released RunLog on typical case.
Previously, only the fail-safe codepath was recording the event.
This change makes it so that the usual codepath records it instead.

This is a quick fix for this issue. Ideally we would refactor the code
so that we only do this in one place.

Bug: 1247227
Change-Id: Ia781b206e6d0b1e10835c5d282e4d9d3074d160d
R=tandrii,yiwzhang
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3142288
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>