[secrets] Add secrets.PrimaryTinkAEAD(ctx), secrets.Encrypt/Decrypt.
The key is initialized via `-primary-tink-aead-key`. It can be used
for various adhoc AEAD operations done by the server.

Requires callers to be careful with providing an adequate context
data in `additionalData`, and thus can be considered a footgun,
since this is not enforceable.

R=tandrii@chromium.org
CC=​weiweilin@google.com

Change-Id: I01c0cf4cbafef104beaefbcacf4f00a344adb5cc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3155686
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>