[cv][api] v0.GetRun returns submission
GetRun doesn't validate data integrity against the stored
run.CLs and run.Submission.

If a CL was listed in both Submission.FailedCLs and
Submission.SucceededCLs, it will be listed in SubmittedCLIndexes only.

If a CL, listed in Submission, was NOT listed in run.CLs,
it will be removed from Submission in the API response.

Bug: 1217300
Change-Id: I2c004bc17183871b5efd43aab3bce5343b5ba888
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3163661
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>