[buildbucket] Extract `fields` mask logic into a separate struct.
It would allow to replace it with an implementation that knows
how to filter structpb.Struct (based on pb.BuildMask).

Note that we don't touch `update_mask` in UpdateBuildRequest, since
it is a different sort of mask. There are some renamings to clarify
that.

R=tandrii@chromium.org
BUG=1219676

Change-Id: Iafd5fc1e33922c5a543272024e8687b488ae9dbf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3166793
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>