[cv] add the owner in the attention set
Gerrit's default rule implies that the owner is automaticallty
added to the attention set when a response is posted in a given CL.

Thus, cancel.Cancel actually skipped adding the owner to the set.
However, I found that it's probably better to add the owner to
the attention set explicitly, because

1) the auto-rule is not visible in unit tests. By adding the owner
explicitly to the attention set, the intent of the unit tests
becomes more clear, when a test validates the resulted attention set.

2) adding the owner to the attention set costs nothing effectively,
but it guarantees that the owner will be added to the attention set
even if the gerrit doc was stale or I completely misunderstood
the rule.

Bug: 1182761
Change-Id: I04eb77b3c7a03d1350912391bba562acfc734a02
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3191395
Auto-Submit: Scott Lee <ddoman@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>