[buildbucket] Add additional fields to global experiment settings.
I anticipate we'll need the builder predicate to do partial rollbacks
as we discover errors when ramping up global experiments (e.g. if a
certain project or builder has problems with the experiment we can
temporarially exclude it but keep it on for the rest of the fleet).

Additionally, the inactive boolean will allow us to affirmatively
distinguish experiment-at-0% from experiment-which-does-nothing
without removing it or keeping a separate list of defunct experiments.

R=chanli, smut, yuanjunh

Bug: 1254504
Change-Id: I737aa56fb7bda09bf31df87aedb2b33c45680c55
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3211681
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>