[tsmon] add missing root labels
Before RootLabels, TargetSchema was used, and it was not necessary
for tsmon to specify all the fields, because they would be set
with the type default values.
e.g., empty string for string fields.

After https://crrev.com/c/3132702, RootLabels are used instead of
TargetSchema and the missing fields are causing data to be dropped
effectively, because the streams would not meet the predicates
in the following recipes - http://shortn/_TkCt56JgKc

All the missing fields in acqusitions.Task were the ones that
ProdX auto adds, if PA endpoint is used. Thus, all the service metrics
weren't affected. However, PopulateProto() for
acqusitions.NetworkDevice was missing some fields in addition to
the auto-added fields, so that bots with newly built sysmon pkg stopped
reporting system metrics since Sep 1, 2021.

This CL adds root labels for the missing fields, so that
reported metric streams will be matched by the predicates.

Tested by running sysmon with this patch, and verified the output
in Monarch.

Bug: 1259823
Change-Id: Idcc0b72d5da7327d6294f9dc84b06e1b32406b49
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3225002
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Michal Kiedys <mkiedys@google.com>