[cv] load exactly `limit` of Runs when querying.
This is important when ACLs and/or other filtering converts
$LIMIT of datastore keys representing Runs to <$LIMIT of actual Runs
to be returned. For practical example, consider ui/recents
which shouldn't show empty list with a page token if the user
has no access to any Runs on the first page.

So, this CL implements continued paging behind the scenes until the
$LIMIT of Runs is reached.

Unfortunately, this isn't perfect: if user can't see *any* Runs,
this will result in a quite expensive and futile search across all of
them, which will likely timeout.

R=robertocn, yiwzhang

Bug: 1264161
Change-Id: I43e0a2c882a51a07eb58510b5e0a22e054a53a20
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3251574
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>