[bbagent] Teach it to run outside of Swarming bot.
When given `-use-gce-account` flag, it would launch a local
LUCI auth server using the GCE metadata server account and use it
for all authentication (in bbagent itself and in subprocesses).

Also populate `realm` section of LUCI_CONTEXT based on the
build proto if it is not populated already.

R=iannucci@chromium.org, tikuta@chromium.org
BUG=1263899

Change-Id: I9b441e1ae137fc03286b6e7ad75953a80cf08e9e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3263221
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>