[buildbucket] Add field to record how Buildbucket determined experiments.
This will let users (and BQ) be able to identify Builds which have an
experiment set as the result of configuration input they didn't
anticipate (e.g.  as a result of the ScheduleBuildRequest or from the
global configuration). This will also allow us to be safer with global
experiment increases because we'll be able to identify Builds which are
in an experiment due to the global experiment change (as opposed to
builder config).

This also removes the previous behavior of setting the bbagent
experiment for Builds running with the non-deprecated `exe` field. The
purpose of setting it previously was to be able to identify "all"
builds using bbagent with a single SearchBuilds query. However this is
the only experiment which behaves in this manner, and the newly added
experiment_reasons field will give substantially better insight.

R=chanli, smut, yuanjunh

Bug: 1254504,1264659
Change-Id: Iec3645dcf740ca1b05ddfb15438e02290fc2ed28
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3271919
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>