[buildbucket] add metrics.WithServiceInfo and WithBuilder
Monitor passes the service info from Server.Init to
ReportBuilderMetrics as a param. It was ok because the call depth
is short. However, such params wouldn't fit well in the RPC protos.

This CL adds WithServiceInfo to store the service info for the entire
life of the server process, and RPCs don't need to pass the info to
metric population functions.

Bug: 1249664
Change-Id: I15e5732641b7952d05b9484b9759114cf195733a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3307575
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: smut <smut@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>