[cv] [tryjob] Model and proto for tryjob watching.
The model is intended to store a minimal association between an
ExternalID and a Tryjob entity that can be retrieved (or determined not
to exist) quickly, when receiving build notifications.

The proto defines the task that would be enqueued by the watcher when a
notification received corresponds to a tryjob that is relevant to one of
our Runs in progress, so that when handled the appropriate RunManager is
notified.

R=yiwzhang,tandrii

Bug: 1257920
Change-Id: I327801e54720b12302c447034ed14bac3eb359e3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3292122
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>