Fix filesystem.Touch behavior with existing files.
Pass O_EXCL to OpenFile so that an error is returned if the file already
exists. This seems to be what the rest of the function is expecting; in
particular it ensures that the file timestamp is still updated to 'now'
if a zero time.Time is passed.

This has been causing vpython to never update venv last-used timestamps
once the venv is created.

Bug: 1276120
Change-Id: Iea210501c8391533929727ac2c2cbdf8c6b1ead8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3325823
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Auto-Submit: Brian Ryner <bryner@google.com>