LogDog: Include charset=utf-8 in the default text content type
Text logs are utf-8 by default, and we want to generally set
the Content-Type header to indicate the encoding.

Content type with alternative encoding may still be given to
logdog_butler, but if not given utf-8 will be used as default.

Content-type for HTML will also be utf-8 by default by can
be overridden to a different text encoding if the logs data
has a different text encoding.

Bug: 1083852
Change-Id: I4962480fe5d47de1f483b586864e8b91e6137ffa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3290301
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>